Clear call related data immediately when leaving call
When a call is left the peer connections are immediately ended. However,
the data of the participants (like the call state), the
CallParticipantModels as well as the timers to retry failed connections
were cleared once the signaling message updating the current participant
state to disconnected was received. If leaving the call failed in the
server that message will not be received, so pending connections were
retried even if no longer in the call, and when joining again a previous
model could be reused, leading to different issues.

Due to that the call related data is now immediately cleared when the
call is left, no matter if the signaling message about leaving the call
is then received or not (and, if it is, nothing will happen as the data
was already cleared).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>