Ignore call signaling messages when not in a call
When a participant was not in the call but received an offer it was
ignored to avoid establishing a connection. However, all the other call
signaling messages were still processed. Due to this, if a participant
leaves the call but that fails in the server that participant still
receives all the participant updates (which also list that participant
as still in the call) and therefore that participant tries to establish
a connection with the other participants.

To prevent that now all call signaling messages are ignored when the
local participant is not in the call. Due to this the call signaling
message that changes the state from the local participant to
disconnected is now ignored as well, although that should not be a
problem given that (since the previous commmit) the call related data is
immediately cleared when the call is left, without waiting for the
signaling message.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>