Fix call left event not sent when leaving conversation while in a call
When a call is left a message is sent to the external signaling server.
However, that message is based on the sessions of the participants in
the conversation. When a conversation is left and the participant was in
a call in that conversation the call is also left, but the event was
emitted after leaving the conversation. Due to that, when the message to
be sent to the external signaling server was generated the participant
was no longer in the conversation, and therefore it was not included in
the message, so the call was not left from the point of view of the
external signaling server. To solve that now the call is left and, once
that is done, the conversation is left.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>