Adjust public share page to layout changes in Nextcloud 25
The content now has an explicit width and height set in the server style
that takes into account the margins, so it does not need to be
overriden.

The footer, on the other hand, now has a fixed position, but as the
element is moved into the #app-content when the Talk sidebar is loaded
it needs to be set as relative instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>