Adjust public share authentication page to layout changes in NC25
The content now has a fixed position to be always centered in the page.
However, in order to add the Talk sidebar the content in the public
share authentication page needs to be "pushed" to the left when the
sidebar is added, so the content position needs to have a relative
position instead (which also keeps the same appearance of the fixed
position when the Talk sidebar is not shown yet).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>