Show error when failing to join a conversation in the public share page
Joining a conversation in the public share page fails if Talk use is
limited to certain groups and the public share page is opened by a user
not in those groups. The incognito mode is enabled when the public share
page is loaded, so from the point of view of the event handler that adds
the Talk sidebar script the user is always a guest, and therefore adding
the script can not be prevented for users not allowed to use Talk.

Moreover, when those users tried to join the conversation in the public
share page the button was kept disabled with a loading spinner forever,
and no error was shown to the user.

For simplicity, and as it could be useful if joining the conversation
fails for any other reason, instead of looking for other ways to hide
the Talk sidebar if the user is not allowed to use Talk now, if joining
the conversation in the public share page fails, the button is restored
and an error toast is shown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>