Ignore messages in other rooms when asserting that a message was sent
When asserting that a message was sent to the signaling server all the
requests were validated before looking for the expected one. As the
requests include the token of the room all the requests were expected to
be sent to the same room; otherwise any request sent to other room would
make the assert fail. Now the requests to other rooms than the room of
the actual message being asserted are ignored, which will make possible
to sent messages to different rooms in the same test.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>