Fix leaving call when switching to another conversation
When a conversation is joined while being in another conversation the
previous conversation is left. Leaving a conversation automatically
leaves the call in the server, but in the WebUI it had no effect on the
call. Therefore, from the point of view of the WebUI the user was still
in the call in the previous conversation, so switching back to it caused
the call view to be shown again (although the call was not actually
joined again, only the call view was shown).

To solve that now an ongoing call is automatically left when its
conversation is left.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>