Fix call view not visible when joining a call again in Files app
When a call is joined all the direct child elements of the header are
hidden by adding a special CSS class, "hidden-by-call", and then the
call view is added to the header. When the call is left the
"hidden-by-call" class is removed and the call view is moved back to
its original parent. That parent was ".app-sidebar-header__action", but
in nextcloud/vue 3.3.0 (Nextcloud 22) it was renamed to
"app-sidebar-header__description".

As the expected parent did not exist the call view was not moved back
and instead stayed as a direct child of the header, so when a call was
joined again "hidden-by-call" was also added to the call view and
therefore hid it like the other elements.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>