Clear store before joining conversation instead of when leaving
When a conversation was left in the Files app the store was cleared to
ensure that it did not contain old data if the conversation was joined
again later (as the conversation data will not be updated while not in
the conversation). However, as this is only needed in case the
conversation is joined again this can be also done before actually
joining the conversation.

This change will be used to automatically leave the call when leaving
the conversation; leaving the call needs to get some values from the
store that were cleared, so this change avoids having to explicitly wait
until the conversation was left before clearing the store.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>