Update room properties with data from "roomlist" update event
The external signaling server includes some room data in the "roomlist"
update event sent when a room is modified. That data is up to date, and
it will be the same received when fetching the room data again, so the
properties can be already updated in the store.

This prevents the lobby from being briefly shown when switching to a
breakout room due to the "switchto" message being handled before the
updated room data could be fetched from the server.

In order to keep the changes to a minimum note that this does was not
applied to guest users, as a different event seems to be sent in that
case, nor to the Talk sidebar, as the current properties provided in the
event should not be relevant to it.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>