Clean breakout room state after stopping them
When the breakout rooms are stopped the lobby is enabled again in them.
However, as it was first enabled and then the breakout rooms were
stopped if the participants updated the room data before switching back
to the parent room the lobby was briefly shown. To prevent that the
breakout rooms should be stopped and, then, the lobby should be enabled
again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>