Add module to start and stop recordings in a call
To record a call a virtual display server and an audio sink are created.
Then a browser is launched inside the virtual display server, and its
audio is routed to the audio sink. This ensures that several browsers
can be running at the same time without interfering with each other.
Then the call is joined by the browser, and an FFMPEG process to record
the virtual display driver and the audio sink is started.

Once the recording is stopped it is uploaded to the Nextcloud server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>