Add helper script to launch the recording backend
The script creates a Docker container and installs all the needed
dependencies as well as the recording backend inside it.

Once the container has been created running the script will just launch
the recording backend again in the existing container, without setting
it up.

Note that the configuration (/etc/nextcloud-talk-recording/server.conf)
needs to be customized as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>