Do not disinvite self-joined users when leaving
If the external signaling server is used, when a participant is
disinvited it is expected that the client leaves the conversation if
still in it. Therefore, self-joined users should not be disinvited when
they disconnect from a conversation, as they can still have other
sessions in it.

To ensure that self-joined users will be disinvited when their last
session leaves the conversation now a USER_REMOVE event is explicitly
triggered when that last session leaves and the user is, in fact,
removed from the conversation.

Note that this does not cause the system message about the participant
leaving to be sent, as it is explicitly guarded against self-joined
users leaving the conversation. However, this change will cause other
handlers to be called now, like those to invalidate resources, but it is
very likely that they should have been called and that this change is
actually a fix too in that regard.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>