Add "starting" values for recording status
As starting the recording can take some time "starting" values were
added to signal that an audio or video recording is starting. The
status will be changed from NONE to STARTING_VIDEO/AUDIO by the
Nextcloud server when the request to start the recording succeeds, and
then the recording status will be set to VIDEO or AUDIO by the recording
server once the recording actually started.

Stopping the recording, on the other hand, should be quite fast, so no
STOPPING status were added.

The system messages were also adjusted to be sent only once the
recording really started or stopped.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>