Handle "starting" values for recording status in UI
The recording indicator in the device checker is still shown without
changes while the recording is starting, as the difference between "is
being recorded" and "will be soon recorded" is not relevant to decide
whether to join or not, or whether to disable the microphone and/or
camera or not. Nevertheless, the condition was made explicit.

On the other hand, the recording indicator in the top bar is now shown
in gray rather than in red while the recording is starting, as the
difference is relevant while in the call to be aware that if something
is said it might not be recorded yet.

The menu action shows a loading spinner and a different text while the
recording is starting. Triggering the menu action while it is starting
stops it, as starting the recording can take a while and this avoids
having to wait if it was wrongly started.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>