Use internal client of signaling server for recording
The recording participant now uses an internal client rather than a
guest participant.

Internal clients are specific to the signaling server and do not have a
counterpart Nextcloud user/session. Authentication is done with a random
string and a token, which is the random string hashed with the internal
secret of the signaling server.

Due to its special nature an internal client is able to join any call,
so using an internal client makes possible to record calls other than
public ones.

Besides that, this also fixes stopping a recording when the last
participant leaves, as the recording participant is no longer a regular
participant known to Nextcloud and therefore does not prevent the call
from ending when the rest left.

Finally, the recording participant is no longer shown in the call either
for the rest of participants, as internal clients are not shown in the
UI (or, at least, they should not be; this may need to be adjusted in
the mobile apps).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>