Do not try to fetch participant data from Nextcloud when recording
The recording participant does not have a Nextcloud session, so it can
not fetch the participant data. Nevertheless, this should be a transient
situation (for example, if a new participant was found but that
participant did not send yet the initial state with the name) and
eventually all the needed data for the call view should be available
from the signaling data.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>