Handle exceptions in async scripts executed through SeleniumHelper
The previous wrapping of the script in an async function swallowed the
exception thrown by the script. If a Promise is passed to the last
argument "execute_async_script" will return the resolved value and throw
the rejected value, so now the async function is, in turn, wrapped by a
Promise which is resolved or rejected as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>