Handle "failed" value for recording status in UI
When a recording failed an error toast is shown, but only for moderators
in the call.

This implies that, depending on the order of signaling messages and how
the conversation data is updated, the message may not be shown if the
recording fails to be stopped when the call is ended for all.
Nevertheless, in that case there will still be a system message in the
chat pointing out that the recording failed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>