Fix adding duplicated breakout rooms to the store
When the breakout rooms returned by the server are processed the
list of breakout rooms for certain parent room may be already
initialized. As the breakout rooms come as an array rather than an
indexed array adding the same breakout room again caused the breakout
room to be duplicated in the list, rather than the previous entry being
overwritten. To solve that now the breakout room to be added is first
removed from the list if it was already there (based on the token).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>