Move packaging related files to its own directory
Instead of using the main build command now a Python package is
explicitly created and extracted, and the Debian package is generated
from that extracted package to be able to adjust some things (like
copying a "setup.py" file used only for packaging without having to
include it in the main sources).

Unfortunately not everything can be done under the "packaging"
directory, as the "__pycache__" and "egg-info" directories are created
in the source directory when the Python package is created.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>