Build Python dependencies not available in repositories
Some Python packages are not available in the Ubuntu repositories, so
now the missing packages are also built along with the recording server
package.

Some dependencies, like Selenium, also have unmet dependencies, so they
are built as well.

The built versions are not the minimum required versions but the latest
versions that could be easily built.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>