Increase default video quality
VP8 encoder now uses the "Constant quality mode" to ensure that every
frame gets the needed bitrate to achieve certain quality level rather
than forcing the whole stream to have an average bitrate.

The maximum bitrate is set to 1Mbit/s, which should be enough to get
clear images even with FullHD screen shares (although this might still
compress some screen shares, as the default bitrate for screen shares in
the HPB is set to 2Mbit/s, but it should be enough for typical screen
shares of slides).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>