Fix RemoteVideoBlocker still active after removing its associated model
Once a CallParticipantModel is removed its associated RemoteVideoBlocker
should no longer block the remote video. Otherwise a deferred blocking
would be rejected by the HPB, as the remote peer is no longer available
or, even worse, it could be accepted and block the video of the remote
peer if it joined again since the original model was removed.

To solve that now the RemoteVideoBlocker is destroyed once its
associated model is removed, which causes the RemoteVideoBlocker to
ignore any further call on its methods. Note that the RemoteVideoBlocker
reference can not be nullified instead on the shared datas, as other
components, like the VideoView, expect a RemoteVideoBlocker to be always
available in its shared data. Internally ignoring further calls also
makes the code more robust than just externally guarding any possible
call.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>