Remove mirroring code when using a shared screen as virtual background
The code might be needed in the original Jitsi project, but not in Talk.
It mirrored the video captured from the camera (and only the video
captured from the camera, the background was not touched), but the local
video is shown mirrored after all the composition, so the local user
would see an unmirrored camera video (due to being double-mirrored) on
top of a mirrored shared screen background, and other participants would
received a mirrored camera video on top of a normal shared screen
background.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>