Add support to use video files as virtual backgrounds
Animated images, like GIFs, can not be used for animated virtual
backgrounds, as only the first frame of the image is taken into account
when drawing an HTMLImageElement to a canvas. To overcome this now
videos can be used too as virtual backgrounds by providing the URL of
the video file, in the same way as with images.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>