Load the virtual background type from the local storage
Similarly to other media settings, like whether audio, video and
virtual background are enabled or not, now the virtual background type
is loaded from the local storage when a call is joined.

Note that video streams are not available, though; blur, image and video
background parameters can be stored in the local storage and reused
between calls, even if Talk was closed and opened again. However, a
video stream is a live object that can not be directly persisted; a key
to trigger capturing a stream again from certain source could be used...
but for now there is no support for persisted video streams.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>