Add signaling messages for reactions in calls
Similar to "raiseHand", reactions are sent in a generic signaling
message of type "message". The reaction is specified in a "reaction"
field with a string value that provides the reaction (which can be any
string, but in general it is expected that a single emoji representing
the reaction will be sent).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>