Fix signaling messages sent to all participants
When a signaling message was sent to all the participants in the call it
was first sent to all the peers and, then, to all the participants that
had no peer. Due to this the signaling message was sent to screen peers
and, when the HPB was used, to the own peer used to send media to the
HPB.

Now those messages only take into account the participants,
independently of the peers related to them, and a single message is sent
for each participant (always with "video" as "roomType", as "roomType"
is always expected in the messages). This prevents, for example,
duplicated messages to be received by a participant sending audio/video
and sharing the screen.

Note that the "sid" parameter is no longer sent in the messages.
Nevertheless, this parameter is specific to identify the session of a
connection, and it is not relevant in the messages sent to all the
participants, so this is not a problem. Moreover, when the HPB is used
its value was useless anyway, as it referred to the connection between
the sender and the HPB, which is unrelated to the receiver connection.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>