Make possible accept PyStringNode as argument
Will be a bit complex to match a string with quote because the regex
ignore quotes: ([^"]*). If we change the regex don't will be possible
found the end of string and maybe we will need to change the delimiter
of the json string:

https://stackoverflow.com/questions/8533943/cucumber-fill-in-a-field-with-double-quote-in-it#8534220

I changed to accept the argument also as PyStringNode and don't need do
nothin more inside the method because the PyStringNode class have a
magich methdo __toString().

Signed-off-by: Vitor Mattos <vitor@php.rio>