fix(guests): Keep guest attendees with display name or non-default permissions
This will allow us to render the display name of guests later on. This was the
idea already since we dropped the talk_guests table in nextcloud/spreed#5146
and we don't delete the attendee anymore on leaving the room since then.
However, the guest clean up when fetching the participant list was still
deleting the guest attendees later on when they didn't have a session.

Signed-off-by: Joas Schilling <coding@schilljs.com>