feat(call): Migrate localStorage settings to Nextcloud's BrowserStorage
Most values are copy-paste, but efore Talk 17 there was only a boolean
`virtualBackgroundEnabled_{token}` (stored as string) for the blur.
Now we also need to have a type and the default type is "none".
So when migrating the data for conversations the user had previously
enabled the background blur we also add the type with value blur.

Signed-off-by: Joas Schilling <coding@schilljs.com>