Fix virtual background image being stretched instead of cropped
If the source values are omitted by default "drawImage" draws the whole
source onto the canvas, therefore stretching it if the background and
the canvas do not have the same aspect ratio. Now the background image
is cropped and centered as needed to cover the whole canvas while
maintaining the original aspect ratio of the background.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>