Fix media settings in files sidebar
When starting a call from the files sidebar and while editing a
file, the media settings was not display because it was covered
by the file dialog. Increasing the z-index of the media settings
dialog didn't fix the issue (it's unclear to me why not).
Removing the container of the media settings dialog and thereby
assigning the default body as container fixes the issue.

Signed-off-by: Marco <marcoambrosini@icloud.com>