Fix trying to use an audio sink as source
In PulseAudio sinks are output devices, like headphones, and playback
streams are written to them, while sources are input devices, like
microphones, and recording streams are read from them.

It is not possible to directly read data from a sink. However, for each
sink a monitor, which is a virtual source, is automatically created,
thus making possible to read the data written to the sink using the
monitor.

The recording server creates a null sink for each browser instance to
write the output audio of the browser to it, but the recorder process
was wrongly given the sink index as its audio input. However, this
happened to work as long as there were no additional sources other than
the monitors created for the null sinks, because in that case the index
of the sinks matched the index of their monitors, and in practice the
recorder process was reading from the monitor of the sink.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>