Add support for Chromium browser
By default Chromium runs in sandboxed mode, which do not seem to work
inside Docker containers (Chromium fails with "Failed to move to new
namespace"); it may work depending on the configuration (for example,
with seccomp profiles), but it does not with the default configuration.

To work around that Chromium is started with "--no-sandbox" by wrapping
the real executable with a script to add that argument.

"wrap_chromium_binary" was adjusted from "NodeChrome/wrap_chrome_binary"
of repository "https://github.com/SeleniumHQ/docker-selenium" at commit
"c6df1ab8dc6a5aca05c163c429a062ada1d79c51".

Docker images for Selenium are licensed under the Apache license 2.0
(see LICENSE.md at above commit).

Besides that, in Selenium 4.6 the Selenium Manager was introduced, which
is a tool that takes care of downloading the browser and its driver if
needed when running Selenium. Unfortunately, when using the "chrome"
driver the Selenium Manager assumes that the executable will be called
"chrome", and if it does not find it, it downloads it to a cache
directory of the current user. However, as the crome driver is found it
will use the one in the path, which may not be compatible with the
latest Chromium version just downloaded, and in the end the version
mismatch causes Chromium to not be launched.

Given that Chromium needs to be launched with "--no-sandbox" and that
the Selenium Manager downloads the executable to its own directory and
launches it directly from there it is not possible either to just leave
the Selenium Manager to install both the browser and its driver.

To work around that the Chromium wrapper is also linked as
"/ur/bin/chrome" to ensure that Selenium Manager will find and use it.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>