Update timestamp only when starting to speak
The timestamp is only used to calculate the elapsed speaking time when
stopping to speak, so it needs to be updated only when starting to
speak.

Additionally, in some cases "speaking" can also be set to "undefined"
(for example, when a remote participant leaves the call); from the point
of view of tracking the speaking time any falsy value is false, so the
value is explicitly set to true/false.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>