Fix speaking status not updated after forced reconnections
When the local participant is force reconnected nothing changes in the
local media model. However, as the ID of the local participant has
changed the speaking status need to be explicitly set to ensure that, if
the participant was speaking, the speaking status will be set to the new
ID (although if this is not explicitly done it will automatically happen
as soon as the participant stops speaking and starts again).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>