Fix getting and setting the virtual background when not available
If the virtual background is not available
"_jitsiStreamBackgroundEffect" is not even defined, so the public getter
and setter can not delegate to it in that case.

As the getter now returns undefined rather than an actual object the
getter can not be used if an object is expected when the virtual
background is not available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>