Log error when capabilities could not be got
If Talkbuchet is not loaded in a working Nextcloud server getting the
capabilities could throw an exception (due to not being able to parse
them), which would abort the load of the rest of the functions. If those
functions are tried to be used later (for example, using the CLI) they
would not be defined and a puzzling error would be shown. Now an
explicit message is logged when the capabilities could not be got, which
gives an explanation to the puzzling error.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>