Set guest name for real participants
Since Talk 18 joining a conversation as a guest requires a name to be
set. If the dialog to set the name is not found it is assumed that an
older Talk version is used and no name is set (even if it would be
possible to do it from the sidebar).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>