Fix "zoom" effect with some cameras when changing the resolution
When the video track resolution is changed the browser may crop and
downscale the camera output to satisfy the requested constraints
("resizeMode: crop-and-scale"). However, in some cameras this causes an
annoying "zoom" effect when changing between resolutions (it might be
caused by the original source being shown for a split second before
being cropped and downscaled, which may change the aspect ratio,
although the reason is not fully clear). To prevent that, now the
browser is asked to provide the camera output as is (although respecting
the given constraints), without further adjustments.

Note that the resize mode is not enforced, though. The browser may still
crop and scale the camera output if it is not possible to get a native
track in the requested resolution range.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>