refactor: Extract playing audio from call participants to its own class
Now the audio elements used to play audio of call participants are all
managed by a single object, rather than each CallParticipantModel
managing the audio elements of a single participant.

Strictly speaking playing audio should be done in the view (as in view
layer, not CallView object, which may or may not exist if Talk is
embedded in Files app and the sidebar is closed), but for simplicity for
now it is kept in the model (layer).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>