fix: Mix audio of call participants in Safari
Safari does not allow autoplaying on audio elements in inactive tabs, so
if the user switched to another tab during a call the participants that
join while the tab is inactive will not be heard until the user switches
back to the Talk tab. To work around that now an audio element is
created when the local participant joins the call and then the audio of
all participants is mixed and played on that audio element.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>