fix: Fix broken audio in Safari after disconnecting node
It seems that Safari loops the audio buffer (or something like that)
when disconnecting an audio node from a destination node being played,
so if a remote participant mutes while speaking Safari plays weird
noises. To work around that now the source of the output media element
is set again whenever a node is disconnected to force creating a new
audio renderer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>