feat: Join and leave federated participants using their federated session
Unfortunately this prevents to join a room as a local federated user, as
in that case the session will be already in use.

On the other hand, this avoids the need to convert sessions between
host and federated servers, and with that also avoids having to keep a
map between sessions.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>