feat: Show active/inactive state of federated participants
Before there was no session for federated participants, so it was not
possible to know if a federated participant was active or not in the
conversation. Now that federated participants join and leave the room
with a session it is possible to show them as active or not in the room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>