refactor: Split server URL from request URL
"baseUrl" and "baseRemoteUrl" are now used only to identify request URLs
(including those from dashboard widgets) while "localServerUrl" and
"remoteServerUrl" are used to identify the local and remote server in
the context of federation. This will be needed to be able to send
requests to different servers while keeping consistent replacement names
for participants.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>