test: Add real federated server for integration tests
Besides the local federation now the server "REMOTE" can be used in
integration tests. The step "using server XXX" can be used to switch
between the LOCAL and REMOTE server, which will change the base URL that
requests are sent to.

The real federated server uses the same code as the main server but a
different configuration and data directory. For simplicity, the
configuration directory is expected to be inside the data directory of
the main server, specifically in
"tests-talk-real-federated-server/config".

If the federated server was not installed in the expected directory
before the tests are run a basic installation using SQLite will be
performed, using "tests-talk-real-federated-server/data" as its data
directory.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>