fix: Provide actor type and id of participants in signaling messages
When the Nextcloud server notifies the signaling server about updates in
the participants it provides a list of properties about each participant
that will be relayed to the clients in the "participants->update"
signaling messages. When federated users were added to the participants
whose properties were relayed the "userId", which until then was set
only for regular users, was set too to the actor id for federated
participants. However, this approach was lacking, as it was not possible
for clients to really know from the "userId" property if the participant
was federated or not or, in a more general way, its actor type.

Due to that the "userId" is again set only for regular users for
backwards compatibility, and the "actorType" and "actorId" properties
are added to the data of each participant included in the
"participants->update" signaling message.

For consistency, "actorType" and "actorId" is also set in the equivalent
message of the internal signaling server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>