fix: Fix calculating stats without enough meaningful values
When the initial stats reports are added the first value of cumulative
stats is used as a base when converting the rest of values to relative.
Therefore, that initial value is not meaningful and should not be used
in calculations, as otherwise the reported quality would be off. Due to
that now no quality is reported until the values of the first report
added were shifted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>