fix: Handle timestamps not updated when stats are stalled
In the past it was observed that the timestamp and round trip time were
updated when the stats stalled (so only the received packets actually
stalled). However, nowadays it seems that the timestamp and round trip
time can also stall. This caused the stalled timestamp to be computed as
0, and that in turn messed with the rest of calculations (for example,
generating a NaN for the number of packets per second) and causing a
"very bad quality" to be wrongly reported. To solve that now the
timestamps are evenly distributed when they unstall.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>