fix: Do not report very bad quality with a low packets count
In the past it was observed that, even for small videos, 10 packets per
second was a reasonable threshold to detect connection issues even if
there were no lost packets. However, nowadays it seems that it can
sometimes trigger a false positive (typically when the background blur
is enabled and the video quality is reduced due to being in a call with
several participants), so for now the connection problem is no longer
reported to the user but just logged.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>