fix: Do not reset previous users after a resume
After the signaling session was resumed the list of known participants
was reset to the next received list of joined users, as it was expected
that the "join" event included the full list of participants currently
in the room. This was introduced when support for the external signaling
server was added in Talk, and back then it may have worked as expected.
However, in current signaling server versions when a "join" event is
received after the session was resumed it only includes the participant
that was just added, rather than the full list (like any other "join"
event except the first one received after joining a room). Therefore all
the previous participants in the room are seen as leaving the room,
which would cause, for example, that they are removed from the call
view. Due to that now the list of participants is no longer reset to the
next received list of joined users.

Moreover, the signaling server also resends any pending message that
would have been received between the session disconnected and resumed,
so the list of known participants will be updated as expected without
that code.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>