fix(performance): Parse file shares inaccurately to prevent setupFS
In some cases we only use the "plain text" version of a chat message.
We can render the conversation list subline without knowing the exact
path of a file. All other information is still available on the cache
entry. The path is only needed in some special cases when afterwards
the viewer is proposed to the user.

Signed-off-by: Joas Schilling <coding@schilljs.com>