fix: Print error code and message for unhandled signaling errors
The whole error was logged, but by default its contents were not visible
unless the object was expanded. Rather than explicitly printing the
error code and message, which are always provided, the whole error is
now stringified. This ensures that any detail will be also readable, as
they would be provided as additional custom attributes.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>