fix: Handle invalid client type when trying to start a recording
If "internalsecret" is not set in the signaling server configuration
internal clients are not supported by the signaling server, and trying
to establish a connection fails with "invalid_client_type". Before, it
was retried again and again to establish a connection until starting the
recording timed out, as no matter how much it was retried it could never
succeed. Due to that now starting the recording fails as soon as an
invalid client type is reported, and a hint about the most likely cause
is provided in the error message printed in the recording server logs.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>