fix(sidebar): mutual events showing incorrect events
Due to a bug in CalDAV search, recurrences that are not a match
will still be returned if the ATTENDEE is found in a recurrence
exception.

See https://github.com/nextcloud/server/issues/52979 for more details
on this bug.

Build an additional filter for the event so we only show mutual events if
the other participant is the organizer or an attendee
of the current recurrence / event.

Signed-off-by: Anna Larch <anna@nextcloud.com>